[Update] 증강 선택 로직 수정 및 선택 동기화, 상태 컨테이너를 딕셔너리로 변경

+++ Assets/Scripts/Weapon/Data/EnhancementDataEntry.cs
@@ -7,7 +7,7 @@ namespace Weapon.Data
         public string Name;
         public string IconName;
         public string Desc;
-        public float Damage;
+        public int Damage;
         public int Magazine;
         public float Delay;
         public float BulletSpd;
